Remove LedgerActions type
We want to execute actions dynamically, without knowing upfront
which fields they will have. For this, we need access to the
Bitcoin & Ethereumwallet.

Since the LedgerActions class is just an indirection through to the
actual wallets, we can flatten the dependency hierachy and depend
on the wallest directly in helloSwap.ts.

This allows us to fetch the necessary data upon constructing a request
from an action.